Core: move IRepl and ReplModule to the client package
