+++ org.metaborg.spoofax.shell.console/src/main/java/org/metaborg/spoofax/shell/client/console/package-info.java
@@ -1,4 +1,4 @@
 /**
- * A console-based implementation of {@link org.metaborg.spoofax.shell.core.IRepl}.
+ * A console-based implementation of {@link org.metaborg.spoofax.shell.client.IRepl}.
  */
 package org.metaborg.spoofax.shell.client.console;
