+++ org.metaborg.spoofax.shell.core/src/main/java/org/metaborg/spoofax/shell/client/IRepl.java
@@ -1,4 +1,4 @@
-package org.metaborg.spoofax.shell.core;
+package org.metaborg.spoofax.shell.client;
 
 import org.metaborg.core.MetaborgException;
 import org.metaborg.spoofax.shell.commands.IReplCommand;
