+++ org.metaborg.spoofax.shell.core/src/main/java/org/metaborg/spoofax/shell/client/hooks/package-info.java
@@ -1,5 +1,5 @@
 /**
- * The {@link org.metaborg.spoofax.shell.core.IRepl} interface uses hooks to pass results back to
+ * The {@link org.metaborg.spoofax.shell.client.IRepl} interface uses hooks to pass results back to
  * the client. This package defines two such hooks.
  */
 package org.metaborg.spoofax.shell.client.hooks;
