+++ org.metaborg.spoofax.shell.core/src/main/java/org/metaborg/spoofax/shell/invoker/package-info.java
@@ -1,5 +1,5 @@
 /**
- * The {@link ICommandInvoker} is called by {@link org.metaborg.spoofax.shell.core.IRepl} to execute
- * expressions and commands.
+ * The {@link ICommandInvoker} is called by {@link org.metaborg.spoofax.shell.client.IRepl} to
+ * execute expressions and commands.
  */
 package org.metaborg.spoofax.shell.invoker;
